Correcciones iniciales
