+++ usuarios/controladores/consultarUsuarios.controlador.js
@@ -1,6 +1,6 @@
 // RF40 Administrador consulta usuarios - https://codeandco-wiki.netlify.app/docs/proyectos/tractores/documentacion/requisitos/RF40
 
-const { consultarUsuarios: consultarUsuariosRepositorio } = require('../data/repositorios/usuarios.repositorio.js');
+const { consultarUsuarios: consultarUsuariosRepositorio } = require('../data/repositorios/modificarUsuarioRepositorio.js.js');
 
 /**
  * Controlador para consultar todos los usuarios
