--- usuarios/data/modelos/usuarios.js
@@ -1,12 +1,0 @@
-const conexion = require('../../../util/servicios/bd.js');
-function consultarRoles() {
-  const consulta = 'SELECT idRol, Nombre FROM rol WHERE idRol != 1';
-  return new Promise((resolve, reject) => {
-    conexion.query(consulta, (error, resultados) => {
-      if (error) return reject(error);
-      resolve(resultados);
-    });
-  });
-}
-
-module.exports = consultarRoles
