+++ usuarios/data/repositorios/crearUsuarioRepositorio.js
@@ -1,0 +1,9 @@
+const modelo = require('../modelos/crearUsuarioModelo.js');
+
+async function crearUsuarioRepositorio(nombre, correo, contrasenia, idRol) {
+    return modelo.crearUsuario(nombre, correo, contrasenia, idRol);
+  }
+
+module.exports = {
+    crearUsuarioRepositorio,
+};
